create function to check if user data on file is accurate

+++ checkout_functions.py
@@ -7,27 +7,10 @@ from datetime import date
 from postage_functions import total_shipping_weight
 import config
 
-
-def checkout():
-    customerdict = {}
-    user_first_name = input("Please enter your first name")
-    user_surname = input("please enter your surname")
-    user_phone_no = input("please enter your phone number")
-    order_date = date.today()
-
-    config.y = user_first_name.lower() + ("_") + user_surname.lower() + (".xml")
-    
-    # write something to check if it's off this format, if they are happy allow it to continue
-    user_address = input("please type your street address with the following format (with commas) - \n number street name, suburb, state , postcode \n for example - 110 Street St, West End, QLD, 4000")
-
-    # Create dictionary with customer's information
-    for i in ('user_first_name', 'user_surname', 'user_phone_no', 'user_address', 'order_date'):
-        customerdict[i] = locals()[i]
-
-    tree_cart = ET.parse("user_cart.xml")
+def user_details_already_in_cart():
+    tree_cart = ET.parse(config.z)
     root_cart = tree_cart.getroot()
 
-    # write customer info to file
     if bool(root_cart.findall("customer", namespaces = None)) == True:
         next
     else: 
@@ -36,17 +19,60 @@ def checkout():
         ET.indent (tree_cart, space = '\t')
 
     customer = root_cart.find("customer")
-    for i, (k, v) in enumerate(customerdict.items()):
-        contact_text = str(v)
-        key_val = str(k)
-        info = ET.SubElement (customer, key_val)
-        info.text = contact_text
-        ET.indent (tree_cart, space = '\t')
-    tree_cart.write("user_cart.xml")
+    for x in customer:
+        if x == None:
+            return False
+        else:  
+            for x in customer:
+                print (x.tag + "----" + x.text)
+            current_details_correct = input("are these still your current details? please type 'yes' or 'no'")
+            if  current_details_correct == "yes":
+                return True
+            else:
+                return False
 
-    file_name = "/Users/nickspringall/Desktop/Coder lessons/terminal_app/user_cart.xml"
-    new_file_name = "/Users/nickspringall/Desktop/Coder lessons/terminal_app/" + (user_first_name.lower() + "_" + user_surname.lower()) + ".xml"
-    os.rename(file_name, new_file_name)
+def checkout():
+
+    if user_details_already_in_cart() == False:
+        customerdict = {}
+        user_first_name = input("Please enter your first name")
+        user_surname = input("please enter your surname")
+        user_phone_no = input("please enter your phone number")
+        oder_date = date.today()
+        user_address = input("please type your street address with the following format (with commas) - \n number street name, suburb, state , postcode \n for example - 110 Street St, West End, QLD, 4000")
+
+ # Create dictionary with customer's information
+        for i in ('user_first_name', 'user_surname', 'user_phone_no', 'user_address', 'order_date'):
+            customerdict[i] = locals()[i]
+            config.y = user_first_name.lower() + ("_") + user_surname.lower() + (".xml")
+    
+        tree_cart = ET.parse(config.y)
+        root_cart = tree_cart.getroot()
+
+    # write customer info to file
+        if bool(root_cart.findall("customer", namespaces = None)) == True:
+            next
+        else: 
+            customer = ET.Element("customer")
+            root_cart.insert(0, customer) 
+            ET.indent (tree_cart, space = '\t')
+
+        customer = root_cart.find("customer")
+        for i, (k, v) in enumerate(customerdict.items()):
+            contact_text = str(v)
+            key_val = str(k)
+            info = ET.SubElement (customer, key_val)
+            info.text = contact_text
+            ET.indent (tree_cart, space = '\t')
+        tree_cart.write("user_cart.xml")
+
+        file_name = "/Users/nickspringall/Desktop/Coder lessons/terminal_app/user_cart.xml"
+        new_file_name = "/Users/nickspringall/Desktop/Coder lessons/terminal_app/" + (user_first_name.lower() + "_" + user_surname.lower()) + ".xml"
+        os.rename(file_name, new_file_name)
+
+    else:
+        config.y = config.z
+
 
 def sub_total(file_name):
     tree_cart = ET.parse(file_name)
@@ -60,10 +86,9 @@ def sub_total(file_name):
 
         quantity = item[0].text
         prod_quant = quantity
-        print(prod_quant)
 
         total_price = float(total_price) + float(float(prod_price_float) * int(prod_quant))
     
-        print("the subtotal of your order is" + str(total_price))
+    print("the subtotal of your order is" + str(total_price))
 
     return total_price
